04-langgraph/fintech_langgraph/requirements.txt
streamlit==1.32.0
langchain>=0.1.0
langchain-community>=0.0.24
langchain-openai>=0.0.2
langchain-chroma>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
sqlalchemy>=2.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
chromadb>=0.4.22
langchain-experimental>=0.0.49
tavily-python>=0.3.0
yfinance>=0.2.36 